Move main.ts to src folder, refactor settings to its own file
